Be resilient to bad queries
With this commit, queries that DuckDB doesn't like
will be rejected, giving the model a chance to try
again.

You can test this with "Compare total_bill quantiles
of smokers and non-smokers", which GPT-4o-mini
reliably fails to do on the first try in my testing. It
seems to be pretty confused about when it can and
cannot use quantile_cont--which to be fair, is pretty
confusing and not that well documented:
https://github.com/observablehq/framework/issues/1092#issuecomment-2002859382